Clean up target frameworks for core library
