+++ System.IO.Abstractions/FileInfoWrapper.cs
@@ -109,7 +109,6 @@ namespace System.IO.Abstractions
             return instance.CreateText();
         }
 
-#if NET40
         public override void Decrypt()
         {
             instance.Decrypt();
@@ -119,7 +118,6 @@ namespace System.IO.Abstractions
         {
             instance.Encrypt();
         }
-#endif
 
         public override FileSecurity GetAccessControl()
         {
@@ -166,7 +164,6 @@ namespace System.IO.Abstractions
             return instance.OpenWrite();
         }
 
-#if NET40
         public override IFileInfo Replace(string destinationFileName, string destinationBackupFileName)
         {
             return new FileInfoWrapper(FileSystem, instance.Replace(destinationFileName, destinationBackupFileName));
@@ -176,7 +173,6 @@ namespace System.IO.Abstractions
         {
             return new FileInfoWrapper(FileSystem, instance.Replace(destinationFileName, destinationBackupFileName, ignoreMetadataErrors));
         }
-#endif
 
         public override void SetAccessControl(FileSecurity fileSecurity)
         {
