+++ System.IO.Abstractions/FileStreamFactory.cs
@@ -8,7 +8,7 @@ namespace System.IO.Abstractions
     {
         public Stream Create(string path, FileMode mode)
             => new FileStream(path, mode);
-    
+
         public Stream Create(string path, FileMode mode, FileAccess access)
             => new FileStream(path, mode, access);
 
@@ -17,13 +17,13 @@ namespace System.IO.Abstractions
 
         public Stream Create(string path, FileMode mode, FileAccess access, FileShare share, int bufferSize)
             => new FileStream(path, mode, access, share, bufferSize);
-    
+
         public Stream Create(string path, FileMode mode, FileAccess access, FileShare share, int bufferSize, FileOptions options)
             => new FileStream(path, mode, access, share, bufferSize, options);
 
         public Stream Create(string path, FileMode mode, FileAccess access, FileShare share, int bufferSize, bool useAsync)
             => new FileStream(path, mode, access, share, bufferSize, useAsync);
-        
+
         public Stream Create(SafeFileHandle handle, FileAccess access)
             => new FileStream(handle, access);
 
@@ -33,15 +33,6 @@ namespace System.IO.Abstractions
         public Stream Create(SafeFileHandle handle, FileAccess access, int bufferSize, bool isAsync)
             => new FileStream(handle, access, bufferSize, isAsync);
 
-#if NET40
-        public Stream Create(string path, FileMode mode, FileSystemRights rights, FileShare share, int bufferSize, FileOptions options, FileSecurity fileSecurity)
-            => new FileStream(path, mode, rights, share, bufferSize, options, fileSecurity);
-
-        public Stream Create(string path, FileMode mode, FileSystemRights rights, FileShare share, int bufferSize, FileOptions options)
-            => new FileStream(path, mode, rights, share, bufferSize, options);
-#endif
-
-#if NET40 || NETSTANDARD_20
         [Obsolete("This method has been deprecated. Please use new Create(SafeFileHandle handle, FileAccess access) instead. http://go.microsoft.com/fwlink/?linkid=14202")]
         public Stream Create(IntPtr handle, FileAccess access)
             => new FileStream(handle, access);
@@ -56,7 +47,6 @@ namespace System.IO.Abstractions
 
         [Obsolete("This method has been deprecated. Please use new Create(SafeFileHandle handle, FileAccess access, int bufferSize, bool isAsync) instead, and optionally make a new SafeFileHandle with ownsHandle=false if needed. http://go.microsoft.com/fwlink/?linkid=14202")]
         public Stream Create(IntPtr handle, FileAccess access, bool ownsHandle, int bufferSize, bool isAsync)
-            =>  new FileStream(handle, access, ownsHandle, bufferSize, isAsync);
-#endif
+            => new FileStream(handle, access, ownsHandle, bufferSize, isAsync);
     }
 }
