+++ System.IO.Abstractions/IFileSystemWatcher.cs
@@ -1,6 +1,4 @@
-﻿#if NET40
-using System.ComponentModel;
-#endif
+﻿using System.ComponentModel;
 
 namespace System.IO.Abstractions
 {
@@ -18,12 +16,10 @@ namespace System.IO.Abstractions
         NotifyFilters NotifyFilter { get; set; }
         /// <inheritdoc cref="FileSystemWatcher.Path"/>
         string Path { get; set; }
-#if NET40
         /// <inheritdoc cref="FileSystemWatcher.Site"/>
         ISite Site { get; set; }
         /// <inheritdoc cref="FileSystemWatcher.SynchronizingObject"/>
         ISynchronizeInvoke SynchronizingObject { get; set; }
-#endif
         /// <inheritdoc cref="FileSystemWatcher.Changed"/>
         event FileSystemEventHandler Changed;
         /// <inheritdoc cref="FileSystemWatcher.Created"/>
@@ -34,14 +30,10 @@ namespace System.IO.Abstractions
         event ErrorEventHandler Error;
         /// <inheritdoc cref="FileSystemWatcher.Renamed"/>
         event RenamedEventHandler Renamed;
-#if NET40
         /// <inheritdoc cref="FileSystemWatcher.BeginInit"/>
         void BeginInit();
-#endif
-#if NET40
         /// <inheritdoc cref="FileSystemWatcher.EndInit"/>
         void EndInit();
-#endif
         /// <inheritdoc cref="FileSystemWatcher.WaitForChanged(WatcherChangeTypes)"/>
         WaitForChangedResult WaitForChanged(WatcherChangeTypes changeType);
         /// <inheritdoc cref="FileSystemWatcher.WaitForChanged(WatcherChangeTypes,int)"/>
