+++ System.IO.Abstractions/PathWrapper.cs
@@ -17,13 +17,11 @@
             get { return Path.DirectorySeparatorChar; }
         }
 
-#if NET40
         [Obsolete("Please use GetInvalidPathChars or GetInvalidFileNameChars instead.")]
         public override char[] InvalidPathChars
         {
             get { return Path.InvalidPathChars; }
         }
-#endif
 
         public override char PathSeparator
         {
