+++ System.IO.Abstractions/System.IO.Abstractions.csproj
@@ -1,59 +1,28 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFrameworks>netstandard1.4;netstandard2.0;netcoreapp2.0;netstandard2.1</TargetFrameworks>
-        <TargetFrameworks Condition="!$([MSBuild]::IsOsUnixLike())">$(TargetFrameworks);net40</TargetFrameworks>
+        <TargetFrameworks>netstandard2.0;netstandard2.1</TargetFrameworks>
         <Authors>Tatham Oddie</Authors>
         <Company />
         <Description>A set of abstractions to help make file system interactions testable.</Description>
         <Copyright>Copyright © Tatham Oddie 2010</Copyright>
-        <AssemblyOriginatorKeyFile>../StrongName.snk</AssemblyOriginatorKeyFile>
         <PackageLicenseUrl>https://github.com/System-IO-Abstractions/System.IO.Abstractions/blob/master/LICENSE</PackageLicenseUrl>
         <PackageProjectUrl>https://github.com/System-IO-Abstractions/System.IO.Abstractions</PackageProjectUrl>
         <PackageTags>testing</PackageTags>
         <GenerateDocumentationFile>true</GenerateDocumentationFile>
         <NoWarn>$(NoWarn);CS1591;CS1574</NoWarn>
+        <SignAssembly Condition="'$(Configuration)' == 'Release'">True</SignAssembly>
+        <AssemblyOriginatorKeyFile>../StrongName.snk</AssemblyOriginatorKeyFile>
+        <DefineConstants Condition="'$(TargetFramework)' == 'netstandard2.0'">$(DefineContstants);</DefineConstants>
+        <DefineConstants Condition="'$(TargetFramework)' == 'netstandard2.1'">$(DefineContstants);FEATURE_ASYNC_FILE</DefineConstants>
     </PropertyGroup>
-
-    <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-        <SignAssembly>False</SignAssembly>
-    </PropertyGroup>
-
-    <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-        <SignAssembly>True</SignAssembly>
-    </PropertyGroup>
-
-    <ItemGroup Condition=" '$(TargetFramework)' == 'netstandard2.0' Or '$(TargetFramework)' == 'netstandard2.1' Or '$(TargetFramework)' == 'netcoreapp2.0'">
-        <PackageReference Include="System.IO.FileSystem.AccessControl">
-            <Version>4.7.0</Version>
-        </PackageReference>
-    </ItemGroup>
-
-    <ItemGroup Condition=" '$(TargetFramework)' == 'netstandard1.4'">
-        <PackageReference Include="System.IO.FileSystem.AccessControl">
-            <Version>4.7.0</Version>
-        </PackageReference>
-        <PackageReference Include="System.IO.FileSystem.DriveInfo">
-            <Version>4.3.1</Version>
-        </PackageReference>
-        <PackageReference Include="System.IO.FileSystem.Watcher">
-            <Version>4.3.0</Version>
-        </PackageReference>
-        <PackageReference Include="System.Runtime.Serialization.Formatters">
-            <Version>4.3.0</Version>
-        </PackageReference>
-    </ItemGroup>
-
-    <ItemGroup Condition=" '$(TargetFramework)' == 'net40' ">
-        <Reference Include="System" />
-        <Reference Include="Microsoft.CSharp" />
-    </ItemGroup>
 
     <ItemGroup>
-      <PackageReference Include="Nerdbank.GitVersioning" Version="3.0.28">
-        <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
-        <PrivateAssets>all</PrivateAssets>
-      </PackageReference>
+        <PackageReference Include="System.IO.FileSystem.AccessControl" Version="4.7.0"/>
+        <PackageReference Include="Nerdbank.GitVersioning" Version="3.0.28">
+            <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
+            <PrivateAssets>all</PrivateAssets>
+        </PackageReference>
     </ItemGroup>
 
 </Project>
